requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
attrs==19.3.0
autopep8==1.5.2
beautifulsoup4==4.9.0
black==19.10b0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==1.1.1
django-compat==1.0.15
django-dotenv==1.4.2
django-extensions==2.2.9
django-ipware==2.1.0
django-rq==2.3.1
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
ipinfo==3.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
redis==3.5.0
regex==2020.4.4
requests==2.23.0
rope==0.16.0
rq==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
ua-parser==0.10.0
urllib3==1.25.9
wrapt==1.12.1
